Dynamic controllers example done.  Also found a bug when routes ended in slash broke things.

+++ Examples/Dynamic Controllers/Examples.DynamicControllers/Resources/ChildResponse.cs
@@ -3,12 +3,12 @@ using APIBlox.AspNetCore.Contracts;
 
 namespace Examples.Resources
 {
-    internal class ChildResponse : IResource<double>
+    internal class ChildResponse : IResource<int>
     {
         public int Age { get; set; }
 
         public string FirstName { get; set; }
-        public double Id { get; set; }
+        public int Id { get; set; }
 
         public string LastName { get; set; }
